+++ src/api/group.ts
@@ -1,4 +1,3 @@
-import request from './request/request'
 import {prepareOptions} from './request/prepareOptions'
 
 /* Types */