+++ src/defaultSettings.ts
@@ -17,6 +17,7 @@ const defaultSettings: DefaultSettingsInterface = {
   multipartChunkSize: 5 * 1024 * 1024, // 5 MB
   multipartMinLastPartSize: 1024 * 1024, // 1MB
   maxConcurrentRequests: 4,
+  pollingTimeoutMilliseconds: 10000,
 }
 
 export default defaultSettings